Reuse output between image and cvideo.
