+++ assets/src/edit-story/elements/image/output.js
@@ -18,47 +18,21 @@
  * Internal dependencies
  */
 import StoryPropTypes from '../../types';
-import { PAGE_WIDTH, PAGE_HEIGHT } from '../../constants';
-import { editorPixels } from '../../units';
-import { getMediaSizePositionProps } from '../media';
+import MediaOutput from '../media/output';
 
 /**
  * Returns AMP HTML for saving into post content for displaying in the FE.
  */
-function ImageOutput({
-  element: { resource, scale, focalX, focalY },
-  box: { width: vw, height: vh },
-}) {
-  // Width and height are taken from the basis of 100% taking into account the
-  // aspect ratio.
-  const width = vw;
-  const height = (vh * PAGE_HEIGHT) / PAGE_WIDTH;
-  const imgProps = getMediaSizePositionProps(
-    resource,
-    width,
-    height,
-    scale,
-    focalX,
-    focalY
-  );
-
-  const wrapperStyle = {
-    position: 'absolute',
-    width: `${editorPixels((imgProps.width / width) * 100)}%`,
-    height: `${editorPixels((imgProps.height / height) * 100)}%`,
-    left: `${-editorPixels((imgProps.offsetX / width) * 100)}%`,
-    top: `${-editorPixels((imgProps.offsetY / height) * 100)}%`,
-  };
-
+function ImageOutput({ element, box }) {
+  const { resource } = element;
   const props = {
     layout: 'fill',
     src: resource.src,
   };
-
   return (
-    <div style={wrapperStyle}>
+    <MediaOutput box={box} element={element}>
       <amp-img {...props} />
-    </div>
+    </MediaOutput>
   );
 }
